# backend/src/generated/swagger.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    AssessmentStatus:
      enum:
        - Present
        - Absent
        - Partial
      type: string
    OralFlag:
      enum:
        - C0
        - C1
        - C2
        - C3
      type: string
...
    DomainScoreRecord:
      properties:
        id:
          type: number
          format: double
        assessmentId:
          type: number
          format: double
        domain:
          $ref: "#/components/schemas/BaselineDomain"
        item1:
          type: number
          format: double
          nullable: true
        item2:
          type: number
          format: double
          nullable: true
        item3:
          type: number
          format: double
          nullable: true
        item4:
          type: number
          format: double
          nullable: true
        item5:
          type: number
          format: double
          nullable: true
        domainScore:
          type: number
          format: double
        suggestedStage:
          $ref: "#/components/schemas/SuggestedStage"
        finalStage:
          allOf:
            - $ref: "#/components/schemas/SuggestedStage"
          nullable: true
        reviewNeeded:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - assessmentId
        - domain
        - item1
        - item2
        - item3
        - item4
        - item5
        - domainScore
        - suggestedStage
        - finalStage
        - reviewNeeded
        - createdAt
      type: object
      additionalProperties: false
    BaselineAssessmentRecord:
      properties:
        id:
          type: number
          format: double
        studentId:
          type: string
        assessmentDate:
          type: string
        assessorName:
          type: string
        status:
          $ref: "#/components/schemas/AssessmentStatus"
        keySupportFlag:
          type: string
          nullable: true
        oralFlag:
          allOf:
            - $ref: "#/components/schemas/OralFlag"
          nullable: true
        qcNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        domainScores:
          items:
            $ref: "#/components/schemas/DomainScoreRecord"
          type: array
      required:
        - id
        - studentId
        - assessmentDate
        - assessorName
        - status
        - keySupportFlag
        - oralFlag
        - qcNotes
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    DomainScoreInput:
      properties:
        domain:
          $ref: "#/components/schemas/BaselineDomain"
        item1:
          type: number
          format: double
          nullable: true
        item2:
          type: number
          format: double
          nullable: true
        item3:
          type: number
          format: double
          nullable: true
        item4:
          type: number
          format: double
          nullable: true
        item5:
          type: number
          format: double
          nullable: true
        domainScore:
          type: number
          format: double
          nullable: true
        suggestedStage:
          allOf:
            - $ref: "#/components/schemas/SuggestedStage"
          nullable: true
        finalStage:
          allOf:
            - $ref: "#/components/schemas/SuggestedStage"
          nullable: true
        reviewNeeded:
          type: boolean
          nullable: true
      required:
        - domain
      type: object
      additionalProperties: false
    CreateBaselineAssessmentInput:
      properties:
        studentId:
          type: string
        assessmentDate:
          type: string
        assessorName:
          type: string
        status:
          $ref: "#/components/schemas/AssessmentStatus"
        keySupportFlag:
          type: string
          nullable: true
        oralFlag:
          allOf:
            - $ref: "#/components/schemas/OralFlag"
          nullable: true
        qcNotes:
          type: string
          nullable: true
        domainScores:
          items:
            $ref: "#/components/schemas/DomainScoreInput"
          type: array
      required:
        - studentId
        - assessmentDate
        - assessorName
        - domainScores
      type: object
      additionalProperties: false
    BulkImportBaselineAssessmentInput:
      properties:
        assessments:
          items:
            $ref: "#/components/schemas/CreateBaselineAssessmentInput"
          type: array
      required:
        - assessments
      type: object
      additionalProperties: false
...
    PaginatedBaselineAssessmentResponse:
      properties:
        records:
          items: {}
          type: array
        total:
          type: number
          format: double
        page:
          type: number
          format: double
        totalPages:
          type: number
          format: double
      required:
        - records
        - total
        - page
        - totalPages
      type: object
      additionalProperties: false
    UpdateBaselineAssessmentInput:
      properties:
        studentId:
          type: string
        assessmentDate:
          type: string
        assessorName:
          type: string
        status:
          $ref: "#/components/schemas/AssessmentStatus"
        keySupportFlag:
          type: string
          nullable: true
        oralFlag:
          allOf:
            - $ref: "#/components/schemas/OralFlag"
          nullable: true
        qcNotes:
          type: string
          nullable: true
        domainScores:
          items:
            $ref: "#/components/schemas/DomainScoreInput"
          type: array
      type: object
      additionalProperties: false
    _36_Enums.AttendanceStatus:
      type: string
      enum:
        - P
        - A
        - HALF_DAY
        - ACTIVITY
        - CANCELLED
    AttendanceStatus:
      $ref: "#/components/schemas/_36_Enums.AttendanceStatus"
    AttendanceStudentSummaryDTO:
      properties:
        id:
          type: number
          format: double
        studentIdCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - id
        - studentIdCode
        - firstName
        - lastName
      type: object
      additionalProperties: false
    AttendanceResponseDTO:
      properties:
        id:
          type: string
        classId:
          type: number
          format: double
        studentId:
          type: number
          format: double
          nullable: true
        sessionDate:
          type: string
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
        className:
          type: string
        student:
          allOf:
            - $ref: "#/components/schemas/AttendanceStudentSummaryDTO"
          nullable: true
      required:
        - id
        - classId
        - studentId
        - sessionDate
        - status
        - remarks
        - createdAt
      type: object
      additionalProperties: false
    BatchUpsertAttendanceResponseDTO:
      properties:
        classId:
          type: number
          format: double
        sessionDate:
          type: string
        upserted:
          type: number
          format: double
        records:
          items:
            $ref: "#/components/schemas/AttendanceResponseDTO"
          type: array
      required:
        - classId
        - sessionDate
        - upserted
        - records
      type: object
      additionalProperties: false
    AttendanceRecordInput:
      properties:
        studentId:
          type: number
          format: double
        status:
          type: string
          enum:
            - P
            - A
            - HALF_DAY
            - ACTIVITY
        remarks:
          type: string
          nullable: true
      required:
        - studentId
        - status
      type: object
      additionalProperties: false
    BatchUpsertAttendanceInput:
      properties:
        classId:
          type: number
          format: double
        sessionDate:
          type: string
        records:
          items:
            $ref: "#/components/schemas/AttendanceRecordInput"
          type: array
      required:
        - classId
        - sessionDate
        - records
      type: object
      additionalProperties: false
    ClassCancellationResponseDTO:
      properties:
        classId:
          type: number
          format: double
        sessionDate:
          type: string
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        remarks:
          type: string
          nullable: true
        removedStudentRecords:
          type: number
          format: double
        record:
          $ref: "#/components/schemas/AttendanceResponseDTO"
      required:
        - classId
        - sessionDate
        - status
        - remarks
        - removedStudentRecords
        - record
      type: object
      additionalProperties: false
    CancelClassAttendanceInput:
      properties:
        classId:
          type: number
          format: double
        sessionDate:
          type: string
        remarks:
          type: string
      required:
        - classId
        - sessionDate
        - remarks
      type: object
      additionalProperties: false
    MutationMessageDTO:
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
      type: object
      additionalProperties: false
    UncancelClassAttendanceInput:
      properties:
        classId:
          type: number
          format: double
        sessionDate:
          type: string
      required:
        - classId
        - sessionDate
      type: object
      additionalProperties: false
...
    DailyRegisterResponseDTO:
      properties:
        classId:
          type: number
          format: double
        className:
          type: string
        sessionDate:
          type: string
        cancelled:
          type: boolean
        cancellationRemarks:
          type: string
          nullable: true
        enrolledCount:
          type: number
          format: double
        markedCount:
          type: number
          format: double
        presentCount:
          type: number
          format: double
        absentCount:
          type: number
          format: double
        halfDayCount:
          type: number
          format: double
        activityCount:
          type: number
          format: double
        entries:
          items:
            $ref: "#/components/schemas/DailyRegisterEntryDTO"
          type: array
      required:
        - classId
        - className
        - sessionDate
        - cancelled
        - cancellationRemarks
        - enrolledCount
        - markedCount
        - presentCount
        - absentCount
        - halfDayCount
        - activityCount
        - entries
      type: object
      additionalProperties: false
    AttendanceRiskLevel:
      enum:
        - STABLE
        - WATCH
        - AT_RISK
        - CRITICAL
      type: string
...
    MonthlyAttendanceAnalyticsDTO:
      properties:
        classId:
          type: number
          format: double
        className:
          type: string
        year:
          type: number
          format: double
        month:
          type: number
          format: double
        workingDays:
          type: number
          format: double
        cancelledDays:
          type: number
          format: double
        classAveragePercent:
          type: number
          format: double
        atRiskCount:
          type: number
          format: double
        students:
          items:
            $ref: "#/components/schemas/StudentMonthlyAnalyticsDTO"
          type: array
      required:
        - classId
        - className
        - year
        - month
        - workingDays
        - cancelledDays
        - classAveragePercent
        - atRiskCount
        - students
      type: object
      additionalProperties: false
    PaginatedAttendanceResponseDTO:
      properties:
        records:
          items:
            $ref: "#/components/schemas/AttendanceResponseDTO"
          type: array
        total:
          type: number
          format: double
        page:
          type: number
          format: double
        limit:
          type: number
          format: double
        totalPages:
          type: number
          format: double
      required:
        - records
        - total
        - page
        - limit
        - totalPages
      type: object
      additionalProperties: false
    UpdateAttendanceInput:
      properties:
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        remarks:
          type: string
          nullable: true
      type: object
      additionalProperties: false
  securitySchemes: {}
info:
  title: lumino1-backend
  version: 1.0.0
  description: Lumino1 NGO Baseline Assessment Production-Ready Express TypeScript Backend
  license:
    name: ISC
  contact:
    name: Edspectrum NGO
paths:
...
  /baseline-assessments:
    post:
      operationId: Create
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineAssessmentRecord"
      description: Create a new Baseline Assessment
      tags:
        - Baseline Assessments
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBaselineAssessmentInput"
    get:
      operationId: List
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBaselineAssessmentResponse"
      description: List paginated Baseline Assessments
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
  /baseline-assessments/import:
    post:
      operationId: ImportAssessments
      responses:
        "201":
          description: Bulk Imported
          content:
            application/json:
              schema:
                properties:
                  records:
                    items:
                      $ref: "#/components/schemas/BaselineAssessmentRecord"
                    type: array
                  importedCount:
                    type: number
                    format: double
                required:
                  - records
                  - importedCount
                type: object
      description: Bulk import Baseline Assessments
      tags:
        - Baseline Assessments
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportBaselineAssessmentInput"
...
  /baseline-assessments/{id}:
    get:
      operationId: GetById
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineAssessmentRecord"
        "404":
          description: Assessment not found
      description: Get Baseline Assessment by numeric ID
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: Update
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineAssessmentRecord"
        "404":
          description: Assessment not found
      description: Update Baseline Assessment by ID
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBaselineAssessmentInput"
    delete:
      operationId: DeleteAssessment
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
        "404":
          description: Assessment not found
      description: Delete Baseline Assessment by ID
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: double
            type: number
  /baseline-assessments/student/{studentId}:
    get:
      operationId: GetByStudentId
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BaselineAssessmentRecord"
                type: array
      description: Get Baseline Assessments by Student ID
      tags:
        - Baseline Assessments
      security: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
  /attendance/batch:
    post:
      operationId: BatchUpsertAttendance
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertAttendanceResponseDTO"
        "400":
          description: Bad Request - Validation Error or Class is Cancelled
        "404":
          description: Class Section not found
      description: Batch mark or update daily attendance records for a class section
      tags:
        - Attendance
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpsertAttendanceInput"
  /attendance/cancel-class:
    post:
      operationId: CancelClassSession
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassCancellationResponseDTO"
        "400":
          description: Bad Request - Validation Error
        "404":
          description: Class Section not found
      description: Cancel an entire class session for a specific date
      tags:
        - Attendance
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelClassAttendanceInput"
  /attendance/uncancel-class:
    post:
      operationId: UncancelClassSession
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationMessageDTO"
        "400":
          description: Bad Request - Validation Error
        "404":
          description: Class Section not found
      description: Remove class session cancellation flag for a specific date
      tags:
        - Attendance
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UncancelClassAttendanceInput"
  /attendance/register:
    get:
      operationId: GetDailyRegister
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyRegisterResponseDTO"
        "400":
          description: Bad Request - Invalid date format
        "404":
          description: Class Section not found
      description: Get daily attendance register grid for a class section and date
      tags:
        - Attendance
      security: []
      parameters:
        - in: query
          name: classId
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: sessionDate
          required: true
          schema:
            type: string
  /attendance/monthly-analytics:
    get:
      operationId: GetMonthlyAnalytics
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyAttendanceAnalyticsDTO"
        "400":
          description: Bad Request - Invalid year/month
        "404":
          description: Class Section not found
      description: Get monthly attendance analytics for a class section
      tags:
        - Attendance
      security: []
      parameters:
        - in: query
          name: classId
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: year
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: month
          required: true
          schema:
            format: double
            type: number
  /attendance/risk-analytics:
    get:
      operationId: GetRiskAnalytics
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyAttendanceAnalyticsDTO"
        "400":
          description: Bad Request - Invalid parameters
        "404":
          description: Class Section not found
      description: Get attendance risk analytics filtered by minimum risk level
      tags:
        - Attendance
      security: []
      parameters:
        - in: query
          name: classId
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: year
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: month
          required: true
          schema:
            format: double
            type: number
        - in: query
          name: minRiskLevel
          required: false
          schema:
            $ref: "#/components/schemas/AttendanceRiskLevel"
  /attendance:
    get:
      operationId: ListAttendance
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAttendanceResponseDTO"
      description: List paginated attendance records with optional filters
      tags:
        - Attendance
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: classId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: studentId
          required: false
          schema:
            format: double
            type: number
        - in: query
          name: sessionDate
          required: false
          schema:
            type: string
        - in: query
          name: fromDate
          required: false
          schema:
            type: string
        - in: query
          name: toDate
          required: false
          schema:
            type: string
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/AttendanceStatus"
  /attendance/{id}:
    get:
      operationId: GetAttendanceById
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponseDTO"
        "404":
          description: Attendance record not found
      description: Get individual attendance record by ID
      tags:
        - Attendance
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    put:
      operationId: UpdateAttendance
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceResponseDTO"
        "404":
          description: Attendance record not found
      description: Update individual attendance record by ID
      tags:
        - Attendance
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceInput"
    delete:
      operationId: DeleteAttendance
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationMessageDTO"
        "404":
          description: Attendance record not found
      description: Delete attendance record by ID
      tags:
        - Attendance
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
servers:
  - url: /